+++ src/index.ts
@@ -2,18 +2,15 @@ export * from "./Session";
 export * from "./SessionManager";
 export * from "./SessionSerializer";
 export * from "./SessionExchanger";
-export {
-  ServerMemoryStorage,
-  serverMemoryStorage,
-  ServerCookieStorage,
-} from "./SessionStorage";
+export { ServerMemoryStorage, ServerCookieStorage } from "./SessionStorage";
 export type {
   SessionStorageInterface,
   ExpirableSessionStorageInterface,
 } from "./SessionStorage";
-export { eventDispatcher } from "./utils";
+export { EventDispatcher } from "./utils";
 export type {
   ListenerFor,
   EventDispatcherInterface,
+  EventDispatcherAware,
   EventRunner,
 } from "./utils";